==> scripts/03_data_analysis/Study 2/01_study2_data_cleaning.ipynb <==
#%pip install davos
import davos

smuggle pandas as pd
smuggle numpy as np
smuggle os

current_directory = os.getcwd()
new_directory = "/Users/suelim/Documents/Research/Evaluation_Vaping_Messages/data/02_Data/Study 2"
os.chdir(new_directory)

df = pd.read_csv("./01_raw_data_anonymized/study2_prolific_anonymized.csv")

# Identify pairs of columns to combine based on the mapping you provided
pairs = [("C_Attention_Check1", "T_Attention_Check1")
] + [("C_Message Rank", "T_Message Ranking")
] + [("C_Message" + str(i), "T_Message" + str(i)) for i in range(1, 6)
] + [
    ("C_Dis_AI" + str(i), "T_Dis_AI" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Dis_Hum" + str(i), "T_Dis_Hum" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Con_AI" + str(i), "T_Con_AI" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Con_Hum" + str(i), "T_Con_Hum" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Unp_AI" + str(i), "T_Unp_AI" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Unp_Hum" + str(i), "T_Unp_Hum" + str(i) + "_1") for i in range(1, 16)
] 

# Create new columns without modifying the original DataFrame in a loop
new_cols = {}
for c_col, t_col in pairs:
    combined_col_name = c_col.replace('C_', '')
    new_cols[combined_col_name] = df[c_col].fillna(df[t_col])

# Concatenate the new columns to the original DataFrame
df = pd.concat([df, pd.DataFrame(new_cols)], axis=1)

df2 = df
cols2 = df2.columns.tolist()
cols2.insert(0, cols2.pop(cols2.index('Experimental_Group')))
df2 = df2[cols2]
df2['Experimental_Group'] = df2['Experimental_Group'].map({0: 'Not Disclosed', 1: 'Disclosed'})
df2["Duration (in seconds)"] = pd.to_numeric(df2["Duration (in seconds)"])
df2 = df2[(df2["Duration (in seconds)"] > 300) & (df2["Finished"] == 1) & (df2["state of residence"] == 22) & (df2["Attention_Check1"] == 2) & (df2["Mani_check_1"] == 5) & (df2["Age"] != "CONSENT_REVOKED")]
# Find the index positions of 'Finished' and 'Age'
Finished_idx = cols2.index('Finished')
Age_idx = cols2.index('Age')
# Reorder the columns
new_cols_order = cols2[:Finished_idx + 1] + cols2[Age_idx:] + cols2[Finished_idx + 1:Age_idx]
# Reassign the columns to the DataFrame based on the new order
df2 = df2[new_cols_order]
df2

df2.to_csv('./02_cleaned_data_file/study2_prolific_cleaned.csv', index=False)

df_sona = pd.read_csv("./01_raw_data_anonymized/study2_sona_anonymized.csv")

cols_sona = df_sona.columns.tolist()

pairs2 = [("C_Message Rank", "T_Message Ranking")
] + [("C_Message" + str(i), "T_Message" + str(i)) for i in range(1, 6)
] + [
    ("C_Dis_AI" + str(i), "T_Dis_AI" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Dis_Hum" + str(i), "T_Dis_Hum" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Con_AI" + str(i), "T_Con_AI" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Con_Hum" + str(i), "T_Con_Hum" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Unp_AI" + str(i), "T_Unp_AI" + str(i) + "_1") for i in range(1, 16)
] + [
    ("C_Unp_Hum" + str(i), "T_Unp_Hum" + str(i) + "_1") for i in range(1, 16)
] 

# Create new columns without modifying the original DataFrame in a loop
new_cols_sona = {}
for c_col, t_col in pairs2:
    combined_col_name = c_col.replace('C_', '')
    new_cols_sona[combined_col_name] = df_sona[c_col].fillna(df_sona[t_col])

# Concatenate the new columns to the original DataFrame
df_sona = pd.concat([df_sona, pd.DataFrame(new_cols_sona)], axis=1)

df_sona2 = df_sona
cols_sona2 = df_sona2.columns.tolist()
cols_sona2.insert(0, cols_sona2.pop(cols_sona2.index('Experimental_Group')))
df_sona2 = df_sona2[cols_sona2]
df_sona2 ['Experimental_Group'] = df_sona2 ['Experimental_Group'].map({0: 'Not Disclosed', 1: 'Disclosed'})
df_sona2["Duration (in seconds)"] = pd.to_numeric(df_sona2["Duration (in seconds)"])
df_sona2 = df_sona2[(df_sona2["Duration (in seconds)"] > 300) & (df_sona2["Finished"] == 1) & (df_sona2["Mani_check_1"] == 5)]
# Find the index positions of 'Finished' and 'Age'
sona_Finished_idx = cols_sona2.index('Finished')
sona_age_idx = cols_sona2.index('Age')
# Reorder the columns
new_cols_order2 = cols_sona2[:sona_Finished_idx + 1] + cols_sona2[sona_age_idx:] + cols_sona2[sona_Finished_idx + 1:sona_age_idx]
# Reassign the columns to the DataFrame based on the new order
df_sona2 = df_sona2[new_cols_order2]
df_sona2

df_sona2.to_csv('./02_cleaned_data_file/study2_sona_cleaned.csv', index=False)

prolific = pd.read_csv("./02_cleaned_data_file/study2_prolific_cleaned.csv")

selected_columns = [0] + list(range(5,7)) + list(range(8, 104)) + list(range(299, 311)) + list(range(312, 320))
prolific_cleaned = prolific.iloc[:, selected_columns]
prolific_cleaned.head()

sona = pd.read_csv("./02_cleaned_data_file/study2_sona_cleaned.csv")

selected_columns2 = [0] + list(range(5, 103)) + list(range(295, 307)) + list(range(308, 316))
sona_cleaned = sona.iloc[:, selected_columns2]
sona_cleaned.head()

combined_df = pd.concat([prolific_cleaned, sona_cleaned], ignore_index=True, sort=False)
combined_df['Participant'] = range(1, len(combined_df) + 1)
cols_comb = combined_df.columns.tolist()
cols_comb.insert(0, cols_comb.pop(cols_comb.index('Participant')))
combined_df = combined_df[cols_comb]
combined_df

combined_df.to_csv('./02_cleaned_data_file/study2_data_file.csv', index=False)
